+++ CHANGES.rst
@@ -28,6 +28,16 @@ Changelog
 - Display positions on the organization view respecting order (getObjPositionInParent).
   [gbastien]
 
+- Display various content title consistently everywhere in the application.
+  [gbastien, sgeulette]
+
+- Display content icon before content title.
+  [gbastien]
+
+- Added parameter display_contact_photo_on_organization_view to the registry,
+  it True (default), the contact photo is displayed in the @@othercontacts, if
+  False, the person content_type icon is displayed.
+  [gbastien]
 
 1.17 (2017-10-02)
 -----------------